requirements.txt
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.133
botocore==1.29.133
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
executing==1.2.0
idna==3.4
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==1.26.15
wcwidth==0.2.6
